fix: moved catalog view to correct folder
